+++ Tests/Unit/CommonResourceHelper.Tests.ps1
@@ -1,0 +1,74 @@
+﻿Describe 'CommonResourceHelper Unit Tests' {
+    BeforeAll {
+        # Import the CommonResourceHelper module to test
+        $dscResourcesFolderFilePath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent) -Parent) `
+                                                -ChildPath 'DscResources'
+
+        Import-Module -Name (Join-Path -Path $dscResourcesFolderFilePath `
+                                       -ChildPath 'CommonResourceHelper.psm1') -Force
+    }
+
+    InModuleScope 'CommonResourceHelper' {
+        $mockTestPath = {
+            return $mockTestPathReturnValue
+        }
+
+        $mockImportLocalizedData = {
+            $BaseDirectory | Should Be $mockExpectedLanguagePath
+        }
+
+        Describe 'Get-LocalizedData' {
+            BeforeEach {
+                Mock -CommandName Test-Path -MockWith $mockTestPath -Verifiable
+                Mock -CommandName Import-LocalizedData -MockWith $mockImportLocalizedData -Verifiable
+            }
+
+            Context 'When loading localized data for Swedish' {
+                $mockExpectedLanguagePath = 'sv-SE'
+                $mockTestPathReturnValue = $true
+
+                It 'Should call Import-LocalizedData with sv-SE language' {
+                    Mock -CommandName Join-Path -MockWith {
+                        return 'sv-SE'
+                    } -Verifiable
+
+                    { Get-LocalizedData 'DummyResource' } | Should Not Throw
+
+                    Assert-MockCalled -CommandName Join-Path -Exactly -Times 2 -Scope It
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 1 -Scope It
+                    Assert-MockCalled -CommandName Import-LocalizedData -Exactly -Times 1 -Scope It
+                }
+
+                $mockExpectedLanguagePath = 'en-US'
+                $mockTestPathReturnValue = $false
+
+                It 'Should call Import-LocalizedData and fallback to en-US if sv-SE language does not exist' {
+                    Mock -CommandName Join-Path -MockWith {
+                        return $ChildPath
+                    } -Verifiable
+
+                    { Get-LocalizedData 'DummyResource' } | Should Not Throw
+
+                    Assert-MockCalled -CommandName Join-Path -Exactly -Times 3 -Scope It
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 1 -Scope It
+                    Assert-MockCalled -CommandName Import-LocalizedData -Exactly -Times 1 -Scope It
+                }
+            }
+
+            Context 'When loading localized data for English' {
+                Mock -CommandName Join-Path -MockWith {
+                    return 'en-US'
+                } -Verifiable
+
+                $mockExpectedLanguagePath = 'en-US'
+                $mockTestPathReturnValue = $true
+
+                It 'Should call Import-LocalizedData with en-US language' {
+                    { Get-LocalizedData 'DummyResource' } | Should Not Throw
+                }
+            }
+
+            Assert-VerifiableMocks
+        }
+    }
+}